+++ Sharely/share_file.py
@@ -5,7 +5,7 @@ from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
 from google.oauth2.credentials import Credentials
 
-import files_information as f_i
+import Sharely.files_information as f_i
 
 SCOPES = ['https://www.googleapis.com/auth/drive']
